fixes #6550 / BZ 1117952 - fix issues with changing org name which affected running db:seed
Without these changes, the user cannot change the default organization
name and re-run 'rake db:seed'.  The reason being, it would attempt
to create an organization with a label that was already in use and
the seed would fail.

In order to address this, this commit does a couple of things:
1. when creating an org, if a label is automatically generated
   and it is already in use, re-generate the label using a uuid.
2. fix an existing validation which was incorrectly comparing
   name=label, label=name, etc. during org create and update.